fastapi-scalable-project/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL异步驱动
aiomysql==0.2.0  # MySQL异步驱动

# 缓存
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 服务发现
python-consul==1.1.0
etcd3==0.12.0

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# 限流和中间件
slowapi==0.1.9
starlette-prometheus==0.9.0

# 任务队列
celery==5.3.4
kombu==5.3.4

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 其他工具
click==8.1.7
rich==13.7.0
typer==0.9.0
